+++ src/AbsolunetTerminal.js
@@ -480,6 +480,25 @@ class AbsolunetTerminal {
 
 
 	/**
+	 * Update the spinner text.
+	 *
+	 * @param {string} text - New text.
+	 * @returns {AbsolunetTerminal} Current instance.
+	 */
+	updateSpinnerText(text) {
+		validateArgument('text', text, requiredStringSchema);
+
+		const spinner = __(this).get('spinner');
+
+		if (typeof spinner !== 'undefined') {
+			spinner.text = text;
+		}
+
+		return this;
+	}
+
+
+	/**
 	 * Stop the spinner.
 	 *
 	 * @returns {AbsolunetTerminal} Current instance.